+++ scripts/restore.pl
@@ -34,8 +34,10 @@ use Kernel::System::ObjectManager;
 
 # get options
 my %Opts;
-my $DB     = '';
-my $DBDump = '';
+my $DB            = '';
+my $DBDump        = '';
+my $DecompressCMD = '';
+my $Installed     = 0;
 getopt( 'hbd', \%Opts );
 if ( exists $Opts{h} ) {
     print <<EOF;
@@ -70,12 +72,39 @@ elsif ( !-d $Opts{d} ) {
     exit 1;
 }
 
+if ( -e File::Spec->catfile( $Opts{b}, 'DatabaseBackup.sql.bz2' ) ) {
+    $DecompressCMD = 'bunzip2';
+}
+else {
+    $DecompressCMD = 'gunzip';
+}
+
+# check needed programs
+for my $CMD ( 'cp', 'tar', $DecompressCMD ) {
+    $Installed = 0;
+    open( my $Input, '-|', "which $CMD" );    ## no critic
+    while (<$Input>) {
+        $Installed = 1;
+    }
+    close $Input;
+    if ( !$Installed ) {
+        print STDERR "ERROR: Can't locate $CMD!\n";
+        exit 1;
+    }
+}
+
 # restore config
-my $ConfigBackupTar = File::Spec->catfile( $Opts{b}, 'Config.tar.gz' );
-print "Restore $ConfigBackupTar ...\n";
 chdir( $Opts{d} );
-if ( -e $ConfigBackupTar ) {
-    system("tar -xzf $ConfigBackupTar");
+
+my $ConfigBackupGz  = File::Spec->catfile( $Opts{b}, 'Config.tar.gz' );
+my $ConfigBackupBz2 = File::Spec->catfile( $Opts{b}, 'Config.tar.bz2' );
+if ( -e $ConfigBackupGz ) {
+    print "Restore $ConfigBackupGz ...\n";
+    system("tar -xzf $ConfigBackupGz");
+}
+elsif ( -e $ConfigBackupBz2 ) {
+    print "Restore $ConfigBackupBz2 ...\n";
+    system("tar -xjf $ConfigBackupBz2");
 }
 
 # create common objects
@@ -101,6 +130,17 @@ if ( $DatabasePw =~ m/^\{(.*)\}$/ ) {
 if ( $DatabaseDSN =~ m/:mysql/i ) {
     $DB     = 'MySQL';
     $DBDump = 'mysql';
+
+    $Installed = 0;
+    open( my $Input, '-|', "which $DBDump" );    ## no critic
+    while (<$Input>) {
+        $Installed = 1;
+    }
+    close $Input;
+    if ( !$Installed ) {
+        print STDERR "ERROR: Can't locate $DBDump!\n";
+        exit 1;
+    }
 }
 elsif ( $DatabaseDSN =~ m/:pg/i ) {
     $DB     = 'PostgreSQL';
@@ -108,24 +148,21 @@ elsif ( $DatabaseDSN =~ m/:pg/i ) {
     if ( $DatabaseDSN !~ m/host=/i ) {
         $DatabaseHost = ''
     }
-}
-else {
-    print STDERR "ERROR: Can't backup, no database dump support!\n";
-    exit 1;
-}
 
-# check needed programs
-for my $CMD ( 'cp', 'tar', $DBDump ) {
-    my $Installed = 0;
-    open( my $Input, '-|', "which $CMD" );    ## no critic
+    $Installed = 0;
+    open( my $Input, '-|', "which $DBDump" );    ## no critic
     while (<$Input>) {
         $Installed = 1;
     }
     close $Input;
     if ( !$Installed ) {
-        print STDERR "ERROR: Can't locate $CMD!\n";
+        print STDERR "ERROR: Can't locate $DBDump!\n";
         exit 1;
     }
+}
+else {
+    print STDERR "ERROR: Can't backup, no database dump support!\n";
+    exit 1;
 }
 
 # check database env
@@ -162,24 +199,39 @@ my $Home = $Kernel::OM->Get('Kernel::Config')->Get('Home');
 chdir($Home);
 
 # extract application
-my $ApplicationBackupTar = File::Spec->catfile( $Opts{b}, 'Application.tar.gz' );
-if ( -e $ApplicationBackupTar ) {
-    print "Restore $ApplicationBackupTar ...\n";
-    system("tar -xzf $ApplicationBackupTar");
+my $ApplicationBackupGz  = File::Spec->catfile( $Opts{b}, 'Application.tar.gz' );
+my $ApplicationBackupBz2 = File::Spec->catfile( $Opts{b}, 'Application.tar.bz2' );
+if ( -e $ApplicationBackupGz ) {
+    print "Restore $ApplicationBackupGz ...\n";
+    system("tar -xzf $ApplicationBackupGz");
+}
+elsif ( -e $ApplicationBackupBz2 ) {
+    print "Restore $ApplicationBackupBz2 ...\n";
+    system("tar -xjf $ApplicationBackupBz2");
 }
 
 # extract vardir
-my $VarDirBackupTar = File::Spec->catfile( $Opts{b}, 'VarDir.tar.gz' );
-if ( -e $VarDirBackupTar ) {
-    print "Restore $VarDirBackupTar ...\n";
-    system("tar -xzf $VarDirBackupTar");
+my $VarDirBackupGz  = File::Spec->catfile( $Opts{b}, 'VarDir.tar.gz' );
+my $VarDirBackupBz2 = File::Spec->catfile( $Opts{b}, 'VarDir.tar.bz2' );
+if ( -e $VarDirBackupGz ) {
+    print "Restore $VarDirBackupGz ...\n";
+    system("tar -xzf $VarDirBackupGz");
+}
+elsif ( -e $VarDirBackupBz2 ) {
+    print "Restore $VarDirBackupBz2 ...\n";
+    system("tar -xjf $VarDirBackupBz2");
 }
 
 # extract datadir
-my $DataDirBackupTar = File::Spec->catfile( $Opts{b}, 'DataDir.tar.gz' );
-if ( -e $DataDirBackupTar ) {
-    print "Restore $DataDirBackupTar ...\n";
-    system("tar -xzf $DataDirBackupTar");
+my $DataDirBackupGz  = File::Spec->catfile( $Opts{b}, 'DataDir.tar.gz' );
+my $DataDirBackupBz2 = File::Spec->catfile( $Opts{b}, 'DataDir.tar.bz2' );
+if ( -e $DataDirBackupGz ) {
+    print "Restore $DataDirBackupGz ...\n";
+    system("tar -xzf $DataDirBackupGz");
+}
+if ( -e $DataDirBackupBz2 ) {
+    print "Restore $DataDirBackupBz2 ...\n";
+    system("tar -xjf $DataDirBackupBz2");
 }
 
 # import database
@@ -191,24 +243,16 @@ if ( $DB =~ m/mysql/i ) {
         $DatabasePw = "-p'$DatabasePw'";
     }
     if ( -e $DatabaseBackupGz ) {
-        print "decompresses SQL-file ...\n";
-        system("gunzip $DatabaseBackupGz");
-        print "cat SQL-file into $DB database\n";
+        print "Restore database into $DB ...\n";
         system(
-            "mysql -f -u$DatabaseUser $DatabasePw -h$DatabaseHost $Database < $Opts{b}/DatabaseBackup.sql"
+            "gunzip -c $DatabaseBackupGz | mysql -f -u$DatabaseUser $DatabasePw -h$DatabaseHost $Database"
         );
-        print "compress SQL-file...\n";
-        system("gzip $Opts{b}/DatabaseBackup.sql");
     }
     elsif ( -e $DatabaseBackupBz2 ) {
-        print "decompresses SQL-file ...\n";
-        system("bunzip2 $DatabaseBackupBz2");
-        print "cat SQL-file into $DB database\n";
+        print "Restore database into $DB ...\n";
         system(
-            "mysql -f -u$DatabaseUser $DatabasePw -h$DatabaseHost $Database < $Opts{b}/DatabaseBackup.sql"
+            "bunzip2 -c $DatabaseBackupBz2 | mysql -f -u$DatabaseUser $DatabasePw -h$DatabaseHost $Database"
         );
-        print "compress SQL-file...\n";
-        system("bzip2 $Opts{b}/DatabaseBackup.sql");
     }
 }
 else {
@@ -217,33 +261,25 @@ else {
     }
 
     if ( -e $DatabaseBackupGz ) {
-        print "decompresses SQL-file ...\n";
-        system("gunzip $DatabaseBackupGz");
 
         # set password via environment variable if there is one
         if ($DatabasePw) {
             $ENV{'PGPASSWORD'} = $DatabasePw;    ## no critic
         }
-        print "cat SQL-file into $DB database\n";
+        print "Restore database into $DB ...\n";
         system(
-            "cat $Opts{b}/DatabaseBackup.sql | psql -U$DatabaseUser $DatabaseHost $Database"
+            "gunzip -c $DatabaseBackupGz | psql -U$DatabaseUser $DatabaseHost $Database"
         );
-        print "compress SQL-file...\n";
-        system("gzip $Opts{b}/DatabaseBackup.sql");
     }
     elsif ( -e $DatabaseBackupBz2 ) {
-        print "decompresses SQL-file ...\n";
-        system("bunzip2 $DatabaseBackupBz2");
 
         # set password via environment variable if there is one
         if ($DatabasePw) {
             $ENV{'PGPASSWORD'} = $DatabasePw;    ## no critic
         }
-        print "cat SQL-file into $DB database\n";
+        print "Restore database into $DB ...\n";
         system(
-            "cat $Opts{b}/DatabaseBackup.sql | psql -U$DatabaseUser $DatabaseHost $Database"
+            "bunzip2 -c $DatabaseBackupBz2 | psql -U$DatabaseUser $DatabaseHost $Database"
         );
-        print "compress SQL-file...\n";
-        system("bzip2 $Opts{b}/DatabaseBackup.sql");
     }
 }